requirements.txt
praw==7.7.1
discord.py==1.7.3